Moved code to seperate helper files
