+++ Sources/ScreenDataUI/Views/SDImage.swift
@@ -7,101 +7,8 @@
 
 import SwiftUI
 import ScreenData
-
 import Combine
 import Task
-
-public protocol SDImageProviding {
-    func image(forURL url: URL) -> AnyPublisher<UIImage?, Error>
-}
-
-public struct SDImageURLProvider: SDImageProviding {
-    public init() { }
-    
-    public func image(forURL url: URL) -> AnyPublisher<UIImage?, Error> {
-        Task.fetch(url: url)
-            .flatMap { (data, response) in
-                Task.do {
-                    guard let data = data else {
-                        return nil
-                    }
-                    let image = UIImage(data: data)
-                    
-                    return image
-                }
-            }
-            .eraseToAnyPublisher()
-    }
-}
-
-public struct SDImageUserDefaultsProvider: SDImageProviding {
-    public init() { }
-    
-    public func image(forURL url: URL) -> AnyPublisher<UIImage?, Error> {
-        Task.do {
-            let imageData: Data? = UserDefaults.standard.data(forKey: url.absoluteString)
-            
-            guard let data = imageData else {
-                return nil
-            }
-            let image = UIImage(data: data)
-            
-            return image
-            
-        }
-        .eraseToAnyPublisher()
-    }
-}
-
-public struct SDImageURLUserDefaultsProvider: SDImageProviding {
-    public init() { }
-    
-    public func image(forURL url: URL) -> AnyPublisher<UIImage?, Error> {
-        SDImageUserDefaultsProvider()
-            .image(forURL: url)
-            .flatMap { image -> AnyPublisher<UIImage?, Error> in
-                guard let image = image else {
-                    return Task.fetch(url: url)
-                        .flatMap { (data, response) in
-                            Task.do {
-                                guard let data = data else {
-                                    return nil
-                                }
-                                let image = UIImage(data: data)
-                                
-                                return image
-                            }
-                        }
-                        .eraseToAnyPublisher()
-                }
-                
-                return Just(image)
-                    .mapError { (Never) -> Error in }
-                    .eraseToAnyPublisher()
-            }
-            .eraseToAnyPublisher()
-        
-    }
-}
-
-public protocol SDImageStoring {
-    func store(image: UIImage?, forURL url: URL) -> AnyPublisher<Void, Error>
-}
-
-public struct SDImageUserDefaultsStorer: SDImageStoring {
-    public init() { }
-    
-    public func store(image: UIImage?, forURL url: URL) -> AnyPublisher<Void, Error> {
-        Task.do {
-            if let imageData = image?.pngData() {
-                UserDefaults.standard.set(imageData, forKey: url.absoluteString)
-            } else {
-                UserDefaults.standard.set(nil, forKey: url.absoluteString)
-            }
-        }
-        .eraseToAnyPublisher()
-    }
-}
 
 public class SDImageStore: ObservableObject {
     @Published public var image: UIImage?
